setup: Loosen dependency version for labgrid
When installing labgridhelper it will first search for an installed
labgrid>=0.1.0. This will fail to find working development versions such
as "0.1.dev1+g3eb7807".

This has a knock-on effect of searching for a suitable "labgrid"
dependency in PyPI. Currently this is labgrid==0.1.0, which contains a
Python 3.7.3 syntax error.

The upshot is that installing labgridhelper with a valid labgrid
version ("0.1.dev1+g3eb7807"), breaks labgrid altogether.

Loosen the dependecy so that working labgrid versions are acceptable
depdenencies.

Signed-off-by: Thomas Preston <thomas.preston@codethink.co.uk>